refactor(bookmarks): inline edit screen into route component
